refactor: move SpinalSimStreamFifoTester
